Tkt 1596 sort resources (#2)
* Refactoring (part I)

* Refactoring (part II)

* Refactoring (part III)

* Refactoring (part IV)

* Refactoring (part V)

* Refactoring (part VI)

* Refactoring (part VII)

* Applying changes
